move kubeconfig utils in separate folder
