--- backend/lib/kubernetes-client/config.js
@@ -1,40 +1,0 @@
-//
-// Copyright (c) 2019 by SAP SE or an SAP affiliate company. All rights reserved. This file is licensed under the Apache Software License, v. 2 except as noted otherwise in the LICENSE file
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//      http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-//
-
-'use strict'
-
-const { getInCluster, fromKubeconfig, readKubeconfig } = require('./util')
-
-function load (env = process.env) {
-  if (/^test/.test(env.NODE_ENV)) {
-    return {
-      url: 'https://kubernetes:6443',
-      auth: {
-        bearer: 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJrdWJlcm5ldGVzL3NlcnZpY2VhY2NvdW50Iiwic3ViIjoic3lzdGVtOnNlcnZpY2VhY2NvdW50OmdhcmRlbjpkZWZhdWx0In0.-4rSuvvj5BStN6DwnmLAaRVbgpl5iCn2hG0pcqx0NPw'
-      }
-    }
-  }
-  if (env.KUBECONFIG) {
-    return fromKubeconfig(readKubeconfig(env.KUBECONFIG))
-  }
-  try {
-    return getInCluster(env)
-  } catch (err) {
-    return fromKubeconfig(readKubeconfig())
-  }
-}
-
-module.exports = load
